Rename to addCoachRoleAction.js
