+++ kolibri/plugins/management/assets/src/state/addCoachRoleAction.js
